requirements.txt
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
